Create text_to_voice base class
